Fix lỗi hiển thị username chậm
