+++ Shareapy/app/src/main/java/com/example/shareapy/models/CurrentUser.java
@@ -1,0 +1,8 @@
+package com.example.shareapy.models;
+
+public class CurrentUser {
+    public static String userID="NULL";
+    public static String userName="NULL";
+
+
+}
